+++ pinbal-scsp/src/main/java/es/caib/pinbal/scsp/Solicitud.java
@@ -23,7 +23,7 @@ public class Solicitud {
 	private String procedimentCodi;
 	private String procedimentNom;
 	private Boolean procedimentValorCampAutomatizado;
-	private Integer procedimentValorCampClaseTramite;
+	private Short procedimentValorCampClaseTramite;
 	private String solicitantIdentificacio;
 	private String solicitantNom;
 	private String funcionariNom;
